spec+proof: make tcb_flag type the same on all arches
In particular, this makes the FpuDisabled flag generic and visible to
all architectures, even ones that do not support FPUs. On those
architectures, the flag is now able to be set but is then ignored by the
rest of the kernel code. This is consistent with the API docs and
behaviour of the C code.

Signed-off-by: Corey Lewis <corey.lewis@proofcraft.systems>